+++ app/lib/shared/versions.dart
@@ -22,10 +22,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2022.08.17',
+  '2022.09.01',
   // Fallback runtime versions.
-  '2022.08.02',
-  '2022.07.05',
+  '2022.08.31',
+  '2022.08.26',
 ];
 
 /// Represents a combined version of the overall toolchain and processing,
@@ -53,11 +53,11 @@ bool shouldGCVersion(String version) =>
     version.compareTo(gcBeforeRuntimeVersion) < 0;
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
-final String runtimeSdkVersion = '2.17.0';
-final String toolStableDartSdkVersion = '2.17.6';
-final String toolStableFlutterSdkVersion = '3.0.5';
-final String toolPreviewDartSdkVersion = '2.18.0-271.4.beta';
-final String toolPreviewFlutterSdkVersion = '3.3.0-0.3.pre';
+final String runtimeSdkVersion = '2.18.0';
+final String toolStableDartSdkVersion = '2.18.0';
+final String toolStableFlutterSdkVersion = '3.3.0';
+final String toolPreviewDartSdkVersion = '2.18.0';
+final String toolPreviewFlutterSdkVersion = '3.3.0';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);